修复sqlite时区问题；修改知识库重新召回逻辑

+++ src/services/files_service.py
@@ -4,6 +4,7 @@ import traceback
 from typing import List
 import sqlite3
 
+from sklearn.metrics import recall_score
 
 status_table_name = 'file_status' # 记录文件信息
 create_status_table_sql = ("create table if not exists file_status "
@@ -118,7 +119,7 @@ class FileStatusManager:
             if self.check_file_exists(file_path, collection_name):
                 return 0
             db.execute(f"insert into {status_table_name} (file_path,collection_name,status,last_updated)"
-                       f" values (?,?,?,datetime('now'))",(file_path,collection_name,status))
+                       f" values (?,?,?,datetime('now', 'localtime'))",(file_path,collection_name,status))
         return 1
 
     def check_file_exists(self,file_path, collection_name):
@@ -130,7 +131,7 @@ class FileStatusManager:
 
     def update_file_status(self,file_path, collection_name, status):
         with SqliteDB(self.db_path) as db:
-            db.execute(f"update {status_table_name} set status = ?,last_updated = datetime('now') where collection_name = ? "
+            db.execute(f"update {status_table_name} set status = ?,last_updated = datetime('now', 'localtime') where collection_name = ? "
                            f"and file_path = ?",(status,collection_name,file_path))
             return db.rowcount
 
@@ -179,6 +180,20 @@ class FileStatusManager:
                        (collection_name, query))
             result = db.fetchone()
             return result[0]
+    def get_search_history_id_by_query(self, collection_name: str, query: str):
+        """
+        获取id
+        :return:
+        """
+        with SqliteDB(self.db_path) as db:
+            db.execute(f"select id from search_history where collection_name = ? and query = ?",
+                       (collection_name, query))
+            result = db.fetchone()
+            if result:
+                return result[0]
+            else:
+                return 0
+
     def add_search_history(self, collection_name: str, query: str, recall_result: List[str], match_best: str):
         """
         添加召回信息
@@ -189,20 +204,43 @@ class FileStatusManager:
             return new_id
         with SqliteDB(self.db_path) as db:
             db.execute(f'insert into search_history (collection_name,query,recall_msg,match_best,create_time)'
-                           f' values (?,?,?,?,datetime("now"))',
+                           f' values (?,?,?,?,datetime("now", "localtime"))',
                            (collection_name, query, json.dumps(recall_result, ensure_ascii=False), match_best))
             db.execute(f'select id from search_history where collection_name = ? and query = ? ',
                        (collection_name, query))
             new_id = db.fetchone()[0]
             return new_id
 
-    def delete_search_history(self, collection_name:str, query: str):
-        history_id = self.check_search_history_exist(collection_name, query)
+    def update_search_history(self, search_id: int, **kwargs):
+        is_update = False
+        update_sql = f"update search_history set  "
+        update_sql_list = []
+        params = []
+        if 'recall_msg' in kwargs:
+            recall_msg = kwargs.get('recall_msg') or ''
+            is_update = True
+            update_sql_list.append(f'recall_msg = ?')
+            params.append(recall_msg)
+        if 'match_best' in kwargs:
+            match_best = kwargs.get('match_best') or ''
+            is_update = True
+            update_sql_list.append(f'match_best = ?')
+            params.append(match_best)
+        if is_update:
+            sql = update_sql + ','.join(update_sql_list) + ' where id = ?'
+            params.append(search_id)
+            with SqliteDB(self.db_path) as db:
+                db.execute(sql, params)
+
+    def delete_search_history(self, collection_name:str, query: str, delete_search=True):
+        history_id = self.get_search_history_id_by_query(collection_name, query)
         if history_id > 0:
             table_id = history_id % 5
             with SqliteDB(self.db_path) as db:
-                db.execute(f'delete from search_history where id = ?', (history_id,))
+                if delete_search:
+                    db.execute(f'delete from search_history where id = ?', (history_id,))
                 db.execute(f'delete from chat_history_{table_id} where search_id = ?', (history_id,))
+        return history_id
 
     def get_collection_search_history(self, collection_name: str, limit: int=1000):
         with SqliteDB(self.db_path) as db:
@@ -227,7 +265,7 @@ class FileStatusManager:
         if search_id > 0:
             with SqliteDB(self.db_path) as db:
                 db.execute(f'insert into chat_history_{search_id % 5} (search_id,chat,create_time)'
-                       f' values (?,?,datetime("now"))',(search_id,chat_text))
+                       f' values (?,?,datetime("now", "localtime"))',(search_id,chat_text))
 
     def get_chat_list(self, search_id: int, page: int=1, page_size: int=100):
         with SqliteDB(self.db_path) as db:
@@ -242,7 +280,7 @@ class FileStatusManager:
             if self.check_file_exists(path, name):
                 return 0
             db.execute(f"insert into base_model_status (name,path,status,create_time)"
-                       f" values (?,?,?,datetime('now'))",(name,path,0))
+                       f" values (?,?,?,datetime('now', 'localtime'))",(name,path,0))
         return 1
 
     def change_base_model(self,name,path):
@@ -253,10 +291,10 @@ class FileStatusManager:
         with SqliteDB(self.db_path) as db:
             if self.check_base_model_exists(name):
                 db.execute(f"update base_model_status set path = ?,status =? ,create_time"
-                           f" = datetime('now') where name = ?",(path,status,name))
+                           f" = datetime('now', 'localtime') where name = ?",(path,status,name))
             else:
                 db.execute(f"insert into base_model_status (name,path,status,create_time)"
-                           f" values (?,?,?,datetime('now'))",(name,path,status))
+                           f" values (?,?,?,datetime('now', 'localtime'))",(name,path,status))
 
     def check_base_model_exists(self, name):
         with SqliteDB(self.db_path) as db:
